Remove colors from Liviano's code
UI has no place in the library, it complicates things and should be only part of the implementation of a wallet, if anything Liviano.CLI  could have colors (but it doesn't do colors).

In order to implement HODLWallet colors once need to store something else, maybe a key of walletId:accountId => colorhex. That way you can render it property.